+++ core/server/api_container/server/api_container_service.go
@@ -26,7 +26,6 @@ import (
 	"github.com/sirupsen/logrus"
 	"google.golang.org/protobuf/types/known/emptypb"
 	"io/ioutil"
-	"math"
 	"net/http"
 	"strings"
 	"time"
@@ -161,19 +160,6 @@ func (apicService ApiContainerService) ExecuteModule(ctx context.Context, args *
 	return resp, nil
 }
 
-func (apicService ApiContainerService) RegisterService(ctx context.Context, args *kurtosis_core_rpc_api_bindings.RegisterServiceArgs) (*kurtosis_core_rpc_api_bindings.RegisterServiceResponse, error) {
-	serviceId := kurtosis_backend_service.ServiceID(args.ServiceId)
-	partitionId := service_network_types.PartitionID(args.PartitionId)
-
-	privateIpAddr, err := apicService.serviceNetwork.RegisterService(ctx, serviceId, partitionId)
-	if err != nil {
-		// TODO IP: Leaks internal information about API container
-		return nil, stacktrace.Propagate(err, "An error occurred registering service '%v' in the service network", serviceId)
-	}
-
-	return binding_constructors.NewRegisterServiceResponse(privateIpAddr.String()), nil
-}
-
 func (apicService ApiContainerService) RegisterServices(ctx context.Context, args *kurtosis_core_rpc_api_bindings.RegisterServicesArgs) (*kurtosis_core_rpc_api_bindings.RegisterServicesResponse, error) {
 	failedServicesPool := map[kurtosis_backend_service.ServiceID]error{}
 	serviceIDs := map[kurtosis_backend_service.ServiceID]bool{}
@@ -219,91 +205,6 @@ func (apicService ApiContainerService) RegisterServices(ctx context.Context, arg
 		delete(shouldRemoveRegistrations, kurtosis_backend_service.ServiceID(serviceIDStr))
 	}
 	return binding_constructors.NewRegisterServicesResponse(serviceIDsToIPsStrs, failedServiceIDsToErrStrs), nil
-}
-
-func (apicService ApiContainerService) StartService(ctx context.Context, args *kurtosis_core_rpc_api_bindings.StartServiceArgs) (*kurtosis_core_rpc_api_bindings.StartServiceResponse, error) {
-	logrus.Debugf("Received request to start service with the following args: %+v", args)
-	serviceId := kurtosis_backend_service.ServiceID(args.ServiceId)
-	privateApiPorts := args.PrivatePorts
-
-	//TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
-	requestedPublicApiPorts := args.PublicPorts
-	if len(requestedPublicApiPorts) > 0 {
-
-		if len(privateApiPorts) != len(requestedPublicApiPorts) {
-			return nil, stacktrace.NewError("The received private ports length and the public ports length are not equal, received '%v' private ports and '%v' public ports", len(privateApiPorts), len(requestedPublicApiPorts))
-		}
-
-		for portId, privatePort := range privateApiPorts {
-			if _, found := requestedPublicApiPorts[portId]; !found {
-				return nil, stacktrace.NewError("Expected to receive public port with ID '%v' bound to private port number '%v', but it was not found", portId, privatePort.GetNumber())
-			}
-		}
-	}
-
-	requestedPublicServicePortSpecs := map[string]*port_spec.PortSpec{}
-	for portId, publicApiPort := range requestedPublicApiPorts {
-		publicServicePortSpec, err := transformApiPortToPortSpec(publicApiPort)
-		if err != nil {
-			return nil, stacktrace.NewError("An error occurred transforming the API port for public port '%v' into a port spec port", portId)
-		}
-		requestedPublicServicePortSpecs[portId] = publicServicePortSpec
-	}
-	//TODO Finished the huge hack to temporarily enable static ports for NEAR
-
-	privateServicePortSpecs := map[string]*port_spec.PortSpec{}
-	for portId, privateApiPort := range privateApiPorts {
-		privateServicePortSpec, err := transformApiPortToPortSpec(privateApiPort)
-		if err != nil {
-			return nil, stacktrace.NewError("An error occurred transforming the API port for private port '%v' into a port spec port", portId)
-		}
-		privateServicePortSpecs[portId] = privateServicePortSpec
-	}
-	filesArtifactMountpointsByArtifactUuid := map[enclave_data_directory.FilesArtifactUUID]string{}
-	for filesArtifactUuidStr, mountDirPath := range args.FilesArtifactMountpoints {
-		filesArtifactMountpointsByArtifactUuid[enclave_data_directory.FilesArtifactUUID(filesArtifactUuidStr)] = mountDirPath
-	}
-	startedService, err := apicService.serviceNetwork.StartService(
-		ctx,
-		serviceId,
-		args.DockerImage,
-		privateServicePortSpecs,
-		requestedPublicServicePortSpecs,
-		args.EntrypointArgs,
-		args.CmdArgs,
-		args.DockerEnvVars,
-		filesArtifactMountpointsByArtifactUuid,
-		args.CpuAllocationMillicpus,
-		args.MemoryAllocationMegabytes,
-	)
-	if err != nil {
-		// TODO IP: Leaks internal information about the API container
-		return nil, stacktrace.Propagate(err, "An error occurred starting the service in the service network")
-	}
-	privateServiceIpStr := startedService.GetRegistration().GetPrivateIP().String()
-	serviceGuidStr := string(startedService.GetRegistration().GetGUID())
-	publicServicePortSpecs := startedService.GetMaybePublicPorts()
-	publicApiPorts, err := transformPortSpecMapToApiPortsMap(publicServicePortSpecs)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred transforming the service's public port specs to API ports")
-	}
-	maybePublicIpAddr := startedService.GetMaybePublicIP()
-	publicIpAddrStr := missingPublicIpAddrStr
-	if maybePublicIpAddr != nil {
-		publicIpAddrStr = maybePublicIpAddr.String()
-	}
-	response := binding_constructors.NewStartServiceResponse(privateServiceIpStr, privateApiPorts, publicIpAddrStr, publicApiPorts, serviceGuidStr)
-
-	serviceStartLoglineSuffix := ""
-	if len(publicServicePortSpecs) > 0 {
-		serviceStartLoglineSuffix = fmt.Sprintf(
-			" with the following public ports: %+v",
-			publicServicePortSpecs,
-		)
-	}
-	logrus.Infof("Started service '%v'%v", serviceId, serviceStartLoglineSuffix)
-
-	return response, nil
 }
 
 func (apicService ApiContainerService) StartServices(ctx context.Context, args *kurtosis_core_rpc_api_bindings.StartServicesArgs) (*kurtosis_core_rpc_api_bindings.StartServicesResponse, error){
@@ -640,33 +541,6 @@ func (apicService ApiContainerService) StoreFilesArtifactFromService(ctx context
 // ====================================================================================================
 // 									   Private helper methods
 // ====================================================================================================
-func transformApiPortToPortSpec(port *kurtosis_core_rpc_api_bindings.Port) (*port_spec.PortSpec, error) {
-	portNumUint32 := port.GetNumber()
-	apiProto := port.GetProtocol()
-	if portNumUint32 > math.MaxUint16 {
-		return nil, stacktrace.NewError(
-			"API port num '%v' is bigger than max allowed port spec port num '%v'",
-			portNumUint32,
-			math.MaxUint16,
-		)
-	}
-	portNumUint16 := uint16(portNumUint32)
-	portSpecProto, found := apiContainerPortProtoToPortSpecPortProto[apiProto]
-	if !found {
-		return nil, stacktrace.NewError("Couldn't find a port spec proto for API port proto '%v'; this should never happen, and is a bug in Kurtosis!", apiProto.String())
-	}
-	result, err := port_spec.NewPortSpec(portNumUint16, portSpecProto)
-	if err != nil {
-		return nil, stacktrace.Propagate(
-			err,
-			"An error occurred creating port spec object with port num '%v' and protocol '%v'",
-			portNumUint16,
-			portSpecProto,
-		)
-	}
-	return result, nil
-}
-
 func transformPortSpecToApiPort(port *port_spec.PortSpec) (*kurtosis_core_rpc_api_bindings.Port, error) {
 	portNumUint16 := port.GetNumber()
 	portSpecProto := port.GetProtocol()